Swapped index files so that signin/login page is first and resource page is second
